Spellshock + Trivium fix (#6430)
- Fix spellshock not counting as magic damage (for trivium/technique)
- Fix Trivium being only labelled "Unscalable Skill"
- Spell Power no longer enforces being float

+++ plugins/paper/src/main/java/com/playmonumenta/plugins/abilities/mage/MagmaShield.java
@@ -113,7 +113,7 @@ public class MagmaShield extends MultipleChargeAbility {
 			return false;
 		}
 
-		float damage = SpellPower.getSpellDamage(mPlugin, mPlayer, (float) (mLevelDamage + mEnhancementDamageBonus * (consumedCharges - 1)));
+		double damage = SpellPower.getSpellDamage(mPlugin, mPlayer, mLevelDamage + mEnhancementDamageBonus * (consumedCharges - 1));
 		double angle = Math.min(CharmManager.calculateFlatAndPercentValue(mPlayer, CHARM_CONE, ANGLE), 180);
 		Hitbox hitbox = Hitbox.approximateCylinderSegment(
 			LocationUtils.getHalfHeightLocation(mPlayer).add(0, -HEIGHT, 0), 2 * HEIGHT, mRadius, Math.toRadians(angle));
